# docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Api App Nội Thất Hoàng Gia
  description: |
    API document for app
    # Base URL:
    http://app-noithat.yez.vn/
    # Status header response code:
    500 - Internal Server Error
    401 - Unauthorized. Token invalid
    200 - OK
    403 - Forbidden
    404 - Not Found


  # host: app-noithat.yez.vn
# schemes:
#    - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
  Token:
    type: apiKey
    name: token
    in: query
...
paths:
...
  /recruitment:
    post:
      security:
        - Bearer: []
      tags:
        - "SP305"
      summary: "recruitment"
      operationId: "Recruitment"
      consumes:
        - "application/x-www-form-urlencoded"
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: id_user
          in: "formData"
          description: "id nhà tuyển dụng"
          required: true
          type: integer
        - name: vacancies
          in: "formData"
          description: "Vị trí tuyển dụng"
          required: true
          type: string
        - name: work_location_name
          in: "formData"
          description: "địa điểm làm việc - address name"
          required: true
          type: array
          items:
            type: string
            format: binary
        - name: work_location_lat_lng
          in: "formData"
          description: "Địa điểm làm việc  - last long"
          required: true
          type: array
          items:
            type: string
            format: binary
        - name: city_name
          in: "formData"
          description: "city_name"
          required: false
          type: string
        - name: service_id[]
          in: "formData"
          description: "Ngành nghề cần tuyển"
          required: true
          type: integer
        - name: wage_min
          in: "formData"
          description: "Mức lương min"
          required: true
          type: array
          items:
            type: string
            format: binary
        - name: wage_max
          in: "formData"
          description: "Mức lương max"
          required: true
          type: array
          items:
            type: string
            format: binary
        - name: number_vacancies
          in: "formData"
          description: "Số lượng cần tuyển"
          required: true
          type: integer
        - name: type
          in: "formData"
          description: "0 : sinh viên | 1: thợ"
          required: true
          type: integer
        - name: experience_id
          in: "formData"
          description: "Yêu cầu kinh nghiệm"
          required: true
          type: integer
        - name: support_mode
          in: "formData"
          description: "Chế độ hỗ trợ"
          required: false
          type: string
        - name: time_work
          in: "formData"
          description: "Thời gian làm việc"
          required: false
          type: string
        - name: description
          in: "formData"
          description: "'Mô tả chi tiết công việc"
          required: false
          type: string
        - name: level_recruitment_id
          in: "formData"
          description: " Id cấp bậc cần tuyển"
          required: false
          type: integer
        - name: probationary_period
          in: "formData"
          description: "Thời gian thử việc"
          required: false
          type: integer
        - name: url_image
          in: "formData"
          description: "Hình ảnh mô tả công việc"
          required: false
          type: array
          items:
            type: string
            format: binary
        - name: url_video
          in: "formData"
          description: "Video mô tả công việc"
          required: false
          type: string
        - name: status
          in: "formData"
          description: " 0: Chưa gửi | 1: Đã gửi sau 1 tuần"
          required: false
          type: integer
      responses:
        200:
          description: Postings job successful
          examples:
            Postings job successful:
              code: 200
              data:
                messages: "Congratulations! you have successfully posted your resume"
                recruiment:
                  id: 10
                  id_user: 30
                  vacancies: "1"
                  work_location_name: "105.8341598"
                  work_location_lat_lng: "21.0277644"
                  city_name: "Ha Noi"
                  wage_min: "3000"
                  wage_max: "6000"
                  number_vacancies: 1
                  type: 1
                  experience_id: 2
                  support_mode: "an, ngu, nghi"
                  time_work: "Từ 8h sáng đến 5h30 chiều"
                  description: "Phụ hồ"
                  level_recruitment_id: 2
                  probationary_period:
                  images:
                    - /images/images/5c0f512790708.png
                    - /images/images/5c0f512790709.png
                    - /images/images/5c0f512790703.png
                  videos: "/videos/video/5c0f51278f6af.mp4"
                  status: 0
                  created_at: "2018-12-11 05:54:47"
                  updated_at: "2018-12-11 05:54:47"
                  service_id:
                    - 1
                    - 2
  /profile:
    post:
      security:
        - Bearer: []
      tags:
        - "SP315"
      summary: "Create new profile"
      operationId: "Profile"
      consumes:
        - "application/x-www-form-urlencoded"
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: user_id
          in: "formData"
          description: "Tên người đăng hồ sơ"
          required: true
          type: integer
        - name: type
          in: "formData"
          description: "Loại hồ sơ 0: sinh viên| 1: thợ thường| 2: thợ nâng cao"
          required: true
          type: integer
        - name: position
          in: "formData"
          description: "Vị trí mong muốn"
          required: true
          type: string
        - name: current_level_id
          in: "formData"
          description: "Cấp bậc hiện tại"
          required: false
          type: integer
        - name: desired_level_id
          in: "formData"
          description: "Cấp bậc mong muốn"
          required: false
          type: integer
        - name: service_id[]
          in: "formData"
          description: "Ngành nghề mong muốn"
          required: false
          type: array
          items:
            type: string
        - name: current_workplace[]
          in: "formData"
          description: "Chỗ làm việc hiện tại"
          required: false
          type: array
          items:
            type: string
        - name: desired_workplace[]
          in: "formData"
          description: "Nơi làm việc mong muốn"
          required: true
          type: array
          items:
            type: string
        - name: experience_id
          in: "formData"
          description: "Số năm kinh nghiệm"
          required: true
          type: integer
        - name: current_salary
          in: "formData"
          description: "Mức lương hiện tại"
          required: false
          type: integer
        - name: desired_salary
          in: "formData"
          description: "Mức lương mong muốn"
          required: true
          type: integer
        - name: skill
          in: "formData"
          description: "Kỹ năng sở trường"
          required: false
          type: string
        - name: "image_product[]"
          in: "formData"
          description: "Hình ảnh sản phẩm đã làm"
          required: false
          type: array
          items:
            type: string
        - name: "image_forte[]"
          in: "formData"
          description: "Sản phẩm sở trường"
          required: false
          type: array
          items:
            type: string
        - name: "image_drawing[]"
          in: "formData"
          description: "Bản vẽ kỹ thuật đã làm "
          required: false
          type: array
          items:
            type: string
        - name: ability_overtime_id
          in: "formData"
          description: "Khả năng tăng ca 0: Dưới 5 buổi/tháng|1: 5-10 buổi/tháng| 2: > 10 buổi/tháng"
          required: false
          type: integer
        - name: ability_manage
          in: "formData"
          description: "Khả năng quản lý nhóm 0: Không|1: Có"
          required: false
          type: integer
        - name: ability_workfar
          in: "formData"
          description: "Khả năng công tác xa  0: Không|1: Có"
          required: false
          type: integer
      responses:
        200:
          description: Create a profile successful
          examples:
            Create a profile successful:
              code: 200
              messages: "Create profile successful"
...
  /profile-detail/{id_profile}:
    get:
      security:
        - Bearer: []
      tags:
        - "SP302"
      summary: "Get profile details"
      operationId: "profile"
      consumes:
        - "application/x-www-form-urlencoded"
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: id_profile
          in: "path"
          description: "id_profile"
          required: true
          type: integer
      responses:
        200:
          description: Get profile details successful
          examples:
            Get profile details successful:
              code: 200
              data:
                messages: "Get profile details successful"
                profiles:
                id: 2
                user_id: 30
                specialize_id: 1
                type: 1
                position: "QA"
                level_id: 2
                career_id: 2
                workplace: "Ha Noi"
                experience_id: 2
                salary: "99999999"
                skill: "quang gach"
                image_product: null
                image_forte: null
                image_drawing: null
                ability_overtime:
                  - id: 1
                    name: Khong
                ability_manage: 1
                ability_workfar: 1
                created_at: null
                updated_at: null
                user:
                  - id_users: 30
                    loaitk: 1
                    vitien: 0
                    name: "Vu Trong Nghiem"
                    username: " 0936788061"
                    email: "nghiemtrongvu01806@gmail.com"
                    telephone: "0936788061"
                    lastvisit: 1536653789
                    active: 1
                    super: 1
                    address: "Ha Noi"
                    address_t: ""
                    address_h: ""
                    address_x: ""
                    dc_nhaxuong: "Nam Hai, Nam Truc"
                    dc_vanphong: "Nháº\u00adp Ä‘á»‹a chá»‰ vÄƒn phÃ²ng cá»§a báº¡n"
                    yahoo: null
                    linkfb: null
                    linktw: null
                    image: "banner-xc6mo0ovwzcpywb-76-doi-tac.jpg"
                    avatar: "55d1cc54687c8822d16d.jpg"
                    cover: null
                    showed: 1
                    linhvucsanxuat: "Ná»™i tháº¥t gá»—"
                    filebaogia: ""
                    idFB: ""
                    idGG: "1"
                    experience: null
                    skill: null
                    code_intro: null
                    code_present: ""
                    service: null
                    identity_card: null
                    type_temp: null
                    point: 0
                    rate: 1
                    gender: 0
                    date_of_birth: "2018-11-20 00:00:00"
                    rate_feedback: 3
                    count_comment: 5
...
host: virtserver.swaggerhub.com
basePath: /thangbeo/register_user/1.0.0
schemes:
  - https
  - http